Add slider images & remove slider component
